+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `IQR` and `RollingIQR` to `stats`
 - The `initializers` submodule has been added to the `optim` module, it can be used for initializing weights in weight-based models such as linear regression
 - `shuffle` to `stream`
+- `Detrender` now has a `window_size` parameter for detrending with a rolling mean
 
 ### Modified